+++ client/src/components/Login-Signup/MenteeSignupForm.js
@@ -1,7 +1,7 @@
 import React from "react";
 import './SignupForm.css';
-import { useFormState } from '../hooks/FormState';
-import { useFormValidation } from '../hooks/FormValidation';
+import { useFormState } from '../../hooks/FormState';
+import { useFormValidation } from '../../hooks/FormValidation';
 import FormField from '../FormField';
 import { menteeFields } from './config/formFields';
