+++ client/src/components/Login-Signup/MentorSignupForm.js
@@ -1,7 +1,7 @@
 import React, {useState} from "react";
 import './SignupForm.css';
-import { useFormState } from '../hooks/FormState';
-import { useFormValidation } from '../hooks/FormValidation';
+import { useFormState } from '../../hooks/FormState';
+import { useFormValidation } from '../../hooks/FormValidation';
 import FormField from '../FormField';
 import { mentorFields } from './config/formFields';
